Re-work documentation
* Update README.md
* Add contributing guidelines summary file: CONTRIBUTING.md
* Add doc8 check on Travis CI -- closes #64
* Add API reference sections for EVM/ACS, triple refinement, ipc, and config.

Add a new structure with sections for:

* overview of the project
* guides, tutorials on how to do specific things
* integrations/applications (e.g.: hyperledger fabric, auction app)
* api/lib reference for the internals (docstrings-based docs)
* contributing/development documentation: how to contribute to the project,
  from getting setup to packaging and distributing the code, passing through
  running the tests, building the docs, working with Travis CI, code coverage,
  git, etc.

The most important thing to keep in mind with respect to this pull request is
that it is simply a basis, to get started with documenting the project. This
means that it can be adapted, re-worked in whatever way is best for the project.